requirements_renderer.txt
Pillow >= 9.5.0